Migrate testing to jest
