+++ package.json
@@ -6,8 +6,8 @@
   "version": "0.4.1",
   "scripts": {
     "start": "webpack-dev-server",
-    "test": "karma start",
-    "test:tdd": "npm run test -- --auto-watch --no-single-run",
+    "test": "jest",
+    "test:tdd": "jest --watch",
     "test:lint": "eslint . --ext .js --ext .jsx --ignore-path .gitignore --ignore-pattern dist --cache",
     "test:ts": "cd tests/typescript && typings install && tsc && rimraf out && rimraf typings",
     "gh-pages": "webpack",
@@ -28,6 +28,7 @@
     "babel-cli": "^6.10.1",
     "babel-core": "^6.10.4",
     "babel-eslint": "^6.1.0",
+    "babel-jest": "15.0.0",
     "babel-loader": "^6.2.4",
     "babel-plugin-istanbul": "^2.0.0",
     "babel-plugin-transform-es2015-destructuring": "~6.9.0",
@@ -36,7 +37,6 @@
     "babel-preset-react": "^6.5.0",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-register": "^6.9.0",
-    "chai": "^3.5.0",
     "clean-webpack-plugin": "^0.1.9",
     "css-loader": "^0.25.0",
     "eslint": "^3.1.0",
@@ -49,23 +49,13 @@
     "highlight.js": "^9.4.0",
     "html-webpack-plugin": "^2.21.0",
     "html-webpack-remark-plugin": "^0.1.0",
+    "jest": "^15.1.1",
     "json-loader": "^0.5.4",
-    "karma": "^1.0.0",
-    "karma-chai": "^0.1.0",
-    "karma-coverage": "^1.0.0",
-    "karma-mocha": "^1.1.1",
-    "karma-phantomjs-launcher": "^1.0.1",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.26",
-    "karma-webpack": "^1.7.0",
-    "mocha": "^3.0.0",
-    "phantomjs-polyfill": "0.0.2",
-    "phantomjs-prebuilt": "^2.1.7",
     "purecss": "^0.6.0",
     "react": "^15.1.0",
-    "react-addons-test-utils": "^15.1.0",
     "react-dom": "^15.1.0",
     "react-ghfork": "^0.3.3",
+    "react-test-renderer": "^15.3.2",
     "rimraf": "^2.5.3",
     "style-loader": "^0.13.1",
     "sync-exec": "^0.6.2",
